+++ src/ledger/SignerRoleHelper.h
@@ -9,6 +9,12 @@ namespace stellar
 class SignerRoleHelper : public EntryHelper
 {
 public:
+    virtual bool
+    exists(LedgerKey const& key) = 0;
+
+    virtual bool
+    exists(uint64_t const roleID) = 0;
+
     virtual SignerRoleFrame::pointer
     loadSignerRole(uint64_t const roleID) = 0;
 };