+++ src/transactions/test/mocks/MockSignerHelper.h
@@ -11,7 +11,6 @@ namespace stellar {
         MOCK_METHOD1(storeChange, void(LedgerEntry const& entry));
         MOCK_METHOD1(storeDelete, void(LedgerKey const& key));
         MOCK_METHOD1(exists, bool(LedgerKey const& key));
-        MOCK_METHOD1(exists, bool(AssetCode const& key));
         MOCK_METHOD1(getLedgerKey, LedgerKey(LedgerEntry const& from));
         MOCK_METHOD1(fromXDR, EntryFrame::pointer(LedgerEntry const& from));
         MOCK_METHOD1(storeLoad, EntryFrame::pointer(LedgerKey const& ledgerKey));
@@ -19,9 +18,12 @@ namespace stellar {
         MOCK_METHOD0(getDatabase, Database&());
         MOCK_METHOD1(flushCachedEntry, void(LedgerKey const& key));
         MOCK_METHOD1(cachedEntryExists, bool(LedgerKey const& key));
+        MOCK_METHOD1(loadSigner, SignerFrame::pointer(PublicKey const& publicKey));
         MOCK_METHOD1(loadSigners,
                      std::vector<SignerFrame::pointer>(AccountID const& accountID));
         MOCK_METHOD1(isRoleIDUsed, bool(uint64_t const roleID));
+        MOCK_METHOD1(exists, bool(PublicKey const& publicKey));
+
 
     };