+++ src/transactions/test/mocks/MockSignerRoleHelper.h
@@ -19,8 +19,9 @@ namespace stellar {
         MOCK_METHOD0(getDatabase, Database&());
         MOCK_METHOD1(flushCachedEntry, void(LedgerKey const& key));
         MOCK_METHOD1(cachedEntryExists, bool(LedgerKey const& key));
-        MOCK_METHOD1(loadSignerRole,
-                SignerRoleFrame::pointer(uint64_t const roleID));
+        MOCK_METHOD1(exists, bool(uint64_t const roleID));
+        MOCK_METHOD1(loadSignerRole, SignerRoleFrame::pointer(uint64_t const roleID));
+
     };
 
 }  // namespace stellar